action fix, avatar loading stuff, fix double register of golf system
